4chan/requirements.txt
requests
pymongo
# faktory
pyfaktory
python-dotenv